requirements.txt
python-telegram-bot==20.6
yt-dlp==2023.11.16
flask==2.3.2
python-dotenv==1.0.0